Static type registration (and separation of said registration) for readability and static access
