+++ Editor/Gui/UiSymbolJson.cs
@@ -21,6 +21,11 @@ namespace T3.Editor.Gui
 {
     public static class SymbolUiJson
     {
+        private static readonly Func<JToken, object> Vector2Reader = JsonToTypeValueConverters.Entries[typeof(Vector2)];
+        private static readonly Func<JToken, object> Vector4Reader = JsonToTypeValueConverters.Entries[typeof(Vector4)];
+        private static readonly Action<JsonTextWriter, object> Vector2Writer = TypeValueToJsonConverters.Entries[typeof(Vector2)];
+        private static readonly Action<JsonTextWriter, object> Vector4Writer = TypeValueToJsonConverters.Entries[typeof(Vector4)];
+        
         public static void WriteSymbolUi(SymbolUi symbolUi, JsonTextWriter writer)
         {
             writer.WriteStartObject();
@@ -65,8 +70,6 @@ namespace T3.Editor.Gui
 
         private static void WriteChildUis(SymbolUi symbolUi, JsonTextWriter writer)
         {
-            var vec2Writer = TypeValueToJsonConverters.Entries[typeof(Vector2)];
-
             writer.WritePropertyName(JsonKeys.SymbolChildUis);
             writer.WriteStartArray();
 
@@ -83,11 +86,11 @@ namespace T3.Editor.Gui
                     if (childUi.Size != SymbolChildUi.DefaultOpSize)
                     {
                         writer.WritePropertyName(JsonKeys.Size);
-                        vec2Writer(writer, childUi.Size);
+                        Vector2Writer(writer, childUi.Size);
                     }
 
                     writer.WritePropertyName(JsonKeys.Position);
-                    vec2Writer(writer, childUi.PosOnCanvas);
+                    Vector2Writer(writer, childUi.PosOnCanvas);
 
                     if (childUi.ConnectionStyleOverrides.Count > 0)
                     {
@@ -112,8 +115,6 @@ namespace T3.Editor.Gui
 
         private static void WriteOutputUis(SymbolUi symbolUi, JsonTextWriter writer)
         {
-            var vec2Writer = TypeValueToJsonConverters.Entries[typeof(Vector2)];
-
             writer.WritePropertyName(JsonKeys.OutputUis);
             writer.WriteStartArray();
 
@@ -125,7 +126,7 @@ namespace T3.Editor.Gui
                 writer.WriteComment(outputName);
                 var outputUi = outputEntry.Value;
                 writer.WritePropertyName(JsonKeys.Position);
-                vec2Writer(writer, outputUi.PosOnCanvas);
+                Vector2Writer(writer, outputUi.PosOnCanvas);
 
                 writer.WriteEndObject();
             }
@@ -137,9 +138,6 @@ namespace T3.Editor.Gui
         {
             if (symbolUi.Annotations.Count == 0)
                 return;
-
-            var vec2Writer = TypeValueToJsonConverters.Entries[typeof(Vector2)];
-            var vec4Writer = TypeValueToJsonConverters.Entries[typeof(Vector4)];
             writer.WritePropertyName(JsonKeys.Annotations);
             writer.WriteStartArray();
 
@@ -150,13 +148,13 @@ namespace T3.Editor.Gui
                 writer.WriteObject(JsonKeys.Title, annotation.Title);
 
                 writer.WritePropertyName(JsonKeys.Color);
-                vec4Writer(writer, annotation.Color.Rgba);
+                Vector4Writer(writer, annotation.Color.Rgba);
 
                 writer.WritePropertyName(JsonKeys.Position);
-                vec2Writer(writer, annotation.PosOnCanvas);
+                Vector2Writer(writer, annotation.PosOnCanvas);
 
                 writer.WritePropertyName(JsonKeys.Size);
-                vec2Writer(writer, annotation.Size);
+                Vector2Writer(writer, annotation.Size);
                 writer.WriteEndObject();
             }
 
@@ -178,11 +176,9 @@ namespace T3.Editor.Gui
             return TryReadSymbolUi(mainObject, symbolId, out symbolUi);
         }
 
+
         internal static bool TryReadSymbolUi(JToken mainObject, Guid symbolId, out SymbolUi symbolUi)
         {
-            var vector2Converter = JsonToTypeValueConverters.Entries[typeof(Vector2)];
-            var vector4Converter = JsonToTypeValueConverters.Entries[typeof(Vector4)];
-
             var symbol = SymbolRegistry.Entries[symbolId];
 
             var inputDict = new OrderedDictionary<Guid, IInputUi>();
@@ -244,12 +240,12 @@ namespace T3.Editor.Gui
                 }
 
                 JToken positionToken = childEntry[JsonKeys.Position];
-                childUi.PosOnCanvas = (Vector2)vector2Converter(positionToken);
+                childUi.PosOnCanvas = (Vector2)Vector2Reader(positionToken);
 
                 if (childEntry[JsonKeys.Size] != null)
                 {
                     JToken sizeToken = childEntry[JsonKeys.Size];
-                    childUi.Size = (Vector2)vector2Converter(sizeToken);
+                    childUi.Size = (Vector2)Vector2Reader(sizeToken);
                 }
 
                 var childStyleEntry = childEntry[JsonKeys.Style];
@@ -304,7 +300,7 @@ namespace T3.Editor.Gui
                     outputUi.OutputDefinition = symbol.OutputDefinitions.First(def => def.Id == outputId);
 
                     JToken positionToken = uiOutputEntry[JsonKeys.Position];
-                    outputUi.PosOnCanvas = (Vector2)vector2Converter(positionToken);
+                    outputUi.PosOnCanvas = (Vector2)Vector2Reader(positionToken);
 
                     outputDict.Add(outputId, outputUi);
                 }
@@ -324,16 +320,16 @@ namespace T3.Editor.Gui
                                          {
                                              Id = Guid.Parse(annotationEntry[JsonKeys.Id].Value<string>()),
                                              Title = annotationEntry[JsonKeys.Title].Value<string>(),
-                                             PosOnCanvas = (Vector2)vector2Converter(annotationEntry[JsonKeys.Position])
+                                             PosOnCanvas = (Vector2)Vector2Reader(annotationEntry[JsonKeys.Position])
                                          };
 
                     var colorEntry = annotationEntry[JsonKeys.Color];
                     if (colorEntry != null)
                     {
-                        annotation.Color = new Color((Vector4)vector4Converter(colorEntry));
+                        annotation.Color = new Color((Vector4)Vector4Reader(colorEntry));
                     }
 
-                    annotation.Size = (Vector2)vector2Converter(annotationEntry[JsonKeys.Size]);
+                    annotation.Size = (Vector2)Vector2Reader(annotationEntry[JsonKeys.Size]);
                     annotationDict[annotation.Id] = annotation;
                 }
             }
